pyproject.toml
[tool.poetry]
name = "evolux-challenge"
version = "0.1.0"
description = "Solution to Evolux challenge"
authors = ["Romildo Oliveira Souza Júnior <mildo@mildo.dev>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "^1.4.0"
flask-jwt-extended = "^4.7.1"
passlib = "^1.7.4"
python-decouple = "^3.8"
psycopg2 = { version = "^2.9.10", optional = true }
gunicorn = { extras = ["gevent"], version = "^23.0.0", optional = true }

[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.10"
coverage = "^7.6.10"
pre-commit = "^4.1.0"

[tool.poetry.extras]
prod = ["psycopg2", "gunicorn"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.coverage.run]
command_line = "-m unittest discover -v tests/"
source = ["src/"]
branch = true

[tool.coverage.report]
fail_under = 100
show_missing = true
